[WIP] Tidy up RemoteServerAuthPage & remember entered credentials
